refactor: move traceroute utils to their own file
